# public/openapi/svc-agent-v1.json
{
  "openapi": "3.1.0",
  "info": {
  "title": "SVC API",
  "version": "1.0.0",
  "description": "Versioned API for role-filtered client, agreement, project, task, time, expense, invoice and received-payment operations. Payment recording is bookkeeping only and never initiates a charge."
},
  "servers": [{"url": "https://svc.bherila.net/api/v1"}],
  "paths": {
    "/workspaces/{workspace_id}/agreements/{agreement_id}/terminate": {"post": {"operationId": "agreements.terminate","security": [{"oauth2": ["clients:write","billing:read"]},{"apiToken": []}],"parameters": [{"$ref": "#/components/parameters/WorkspaceId"},{"name": "agreement_id","in": "path","required": true,"schema": {"type": "string","format": "uuid"}},{"$ref": "#/components/parameters/IdempotencyKey"}],"responses": {"200": {"description": "Authorized result","content": {"application/json": {"schema": {"$ref": "#/components/schemas/AgreementResponse"}}}},"default": {"$ref": "#/components/responses/Error"}},"requestBody": {"required": true,"content": {"application/json": {"schema": {"$ref": "#/components/schemas/agreements_terminateRequest"}}}},"description": "Workspace managers only. Requires both AGENT_API_WRITES_ENABLED and AGENT_API_CLIENT_WRITES_ENABLED. Versioned writes also require the corresponding read scope. Enabling automatic delivery requires a delay and a valid billing email or client portal recipient. Existing records require their current opaque expected_version; agreement creation uses the parent client version. Activation and termination require confirm: true.","x-mcp-tool": "agreements.terminate","x-svc-flags": ["AGENT_API_WRITES_ENABLED","AGENT_API_CLIENT_WRITES_ENABLED"]}},
    "/workspaces/{workspace_id}/agreements/{agreement_id}/activate": {"post": {"operationId": "agreements.activate","security": [{"oauth2": ["clients:write","billing:read"]},{"apiToken": []}],"parameters": [{"$ref": "#/components/parameters/WorkspaceId"},{"name": "agreement_id","in": "path","required": true,"schema": {"type": "string","format": "uuid"}},{"$ref": "#/components/parameters/IdempotencyKey"}],"responses": {"200": {"description": "Authorized result","content": {"application/json": {"schema": {"$ref": "#/components/schemas/AgreementResponse"}}}},"default": {"$ref": "#/components/responses/Error"}},"requestBody": {"required": true,"content": {"application/json": {"schema": {"$ref": "#/components/schemas/agreements_activateRequest"}}}},"description": "Workspace managers only. Requires both AGENT_API_WRITES_ENABLED and AGENT_API_CLIENT_WRITES_ENABLED. Versioned writes also require the corresponding read scope. Enabling automatic delivery requires a delay and a valid billing email or client portal recipient. Existing records require their current opaque expected_version; agreement creation uses the parent client version. Activation and termination require confirm: true.","x-mcp-tool": "agreements.activate","x-svc-flags": ["AGENT_API_WRITES_ENABLED","AGENT_API_CLIENT_WRITES_ENABLED"]}},
...
    "/workspaces/{workspace_id}/clients/{client_id}/restore": {"post": {"operationId": "clients.restore","security": [{"oauth2": ["clients:write","clients:read"]},{"apiToken": []}],"parameters": [{"$ref": "#/components/parameters/WorkspaceId"},{"name": "client_id","in": "path","required": true,"schema": {"type": "string","format": "uuid"}},{"$ref": "#/components/parameters/IdempotencyKey"}],"responses": {"200": {"description": "Authorized result","content": {"application/json": {"schema": {"$ref": "#/components/schemas/ClientCompanyResponse"}}}},"default": {"$ref": "#/components/responses/Error"}},"requestBody": {"required": true,"content": {"application/json": {"schema": {"$ref": "#/components/schemas/clients_restoreRequest"}}}},"description": "Workspace managers only. Requires both AGENT_API_WRITES_ENABLED and AGENT_API_CLIENT_WRITES_ENABLED. Versioned writes also require the corresponding read scope. Enabling automatic delivery requires a delay and a valid billing email or client portal recipient. Existing records require their current opaque expected_version; agreement creation uses the parent client version. Activation and termination require confirm: true.","x-mcp-tool": "clients.restore","x-svc-flags": ["AGENT_API_WRITES_ENABLED","AGENT_API_CLIENT_WRITES_ENABLED"]}},
    "/workspaces/{workspace_id}/clients/{client_id}/archive": {"post": {"operationId": "clients.archive","security": [{"oauth2": ["clients:write","clients:read"]},{"apiToken": []}],"parameters": [{"$ref": "#/components/parameters/WorkspaceId"},{"name": "client_id","in": "path","required": true,"schema": {"type": "string","format": "uuid"}},{"$ref": "#/components/parameters/IdempotencyKey"}],"responses": {"200": {"description": "Authorized result","content": {"application/json": {"schema": {"$ref": "#/components/schemas/ClientCompanyResponse"}}}},"default": {"$ref": "#/components/responses/Error"}},"requestBody": {"required": true,"content": {"application/json": {"schema": {"$ref": "#/components/schemas/clients_archiveRequest"}}}},"description": "Workspace managers only. Requires both AGENT_API_WRITES_ENABLED and AGENT_API_CLIENT_WRITES_ENABLED. Versioned writes also require the corresponding read scope. Enabling automatic delivery requires a delay and a valid billing email or client portal recipient. Existing records require their current opaque expected_version; agreement creation uses the parent client version. Activation and termination require confirm: true.","x-mcp-tool": "clients.archive","x-svc-flags": ["AGENT_API_WRITES_ENABLED","AGENT_API_CLIENT_WRITES_ENABLED"]}},
...
  },
  "components": {
    "securitySchemes": {"apiToken": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "A personal API token created on the SVC setup page. It carries the scopes chosen when it was created and expires after the chosen lifetime."}, "oauth2": {
      "type": "oauth2", "description": "Authorization code with S256 PKCE and interactive Bherila.net login.",
      "flows": {"authorizationCode": {
        "authorizationUrl": "https://svc.bherila.net/oauth/authorize", "tokenUrl": "https://svc.bherila.net/oauth/token",
        "scopes": {
          "expenses:read": "Read authorized expenses", "expenses:write": "Record, edit and delete draft expenses as a workspace manager", "mcp:use": "Connect through MCP", "identity:read": "Read identity and workspaces", "projects:read": "Read projects",
          "tasks:read": "Read tasks", "tasks:write": "Create and update tasks", "time:read": "Read time",
          "time:write": "Log and manage draft time", "time:approve": "Set billing rates and approve project time", "billing:read": "Read invoices",
          "billing:write": "Create and edit invoice drafts", "payments:read": "Read received payments on authorized invoices", "payments:record": "Record money already received; never initiate a charge", "billing:deliver": "Issue, send, correct, and void invoices",
          "clients:read": "Read client companies as a workspace manager", "clients:write": "Create, edit and archive client companies and their agreements as a workspace manager"
        }
      }}
    }},
    "parameters": {
...
      "WorkspaceId": {"name": "workspace_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
...
      "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 255}}
    },
    "responses": {"Error": {"description": "OAuth, authorization, tenant, conflict, validation, or rate-limit error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}},
    "schemas": {
      "agreements_terminateRequest": {"type": "object","additionalProperties": false,"required": ["confirm","expected_version"],"properties": {"confirm": {"type": "boolean","const": true},"expected_version": {"type": "string","minLength": 64,"maxLength": 64},"ends_on": {"type": ["string","null"],"format": "date"}}},
      "agreements_activateRequest": {"type": "object","additionalProperties": false,"required": ["confirm","expected_version"],"properties": {"confirm": {"type": "boolean","const": true},"expected_version": {"type": "string","minLength": 64,"maxLength": 64}}},
...
      "clients_restoreRequest": {"type": "object","additionalProperties": false,"required": ["expected_version"],"properties": {"expected_version": {"type": "string","minLength": 64,"maxLength": 64}}},
      "clients_archiveRequest": {"type": "object","additionalProperties": false,"required": ["expected_version"],"properties": {"expected_version": {"type": "string","minLength": 64,"maxLength": 64}}},
...
      "AgreementResponse": {"type": "object","additionalProperties": false,"required": ["data"],"properties": {"data": {"$ref": "#/components/schemas/Agreement"}}},
...
      "ClientCompanyResponse": {"type": "object","additionalProperties": false,"required": ["data"],"properties": {"data": {"$ref": "#/components/schemas/ClientCompany"}}},
      "Agreement": {"type": "object","additionalProperties": false,"required": ["version","id","title","status","currency","billing_cadence","effective_billing_cadence","effective_first_cycle_proration","is_recurring","starts_on","ends_on","signed_at","retainer_minutes_per_period","retainer_minutes_per_month","retainer_amount_per_period","hourly_rate_amount","rollover_months","project","client_id","client_name","project_id"],"properties": {"version": {"type": "string","minLength": 64,"maxLength": 64},"id": {"type": "string","format": "uuid"},"title": {"type": "string","maxLength": 255},"status": {"type": "string","maxLength": 64},"currency": {"type": "string","minLength": 3,"maxLength": 3},"billing_cadence": {"type": ["string","null"],"maxLength": 32},"effective_billing_cadence": {"type": ["string","null"],"enum": ["monthly","quarterly","semi_annual","annual",null]},"effective_first_cycle_proration": {"type": ["string","null"],"enum": ["prorate_hours","full_period","align_next_cycle",null]},"is_recurring": {"type": "boolean"},"starts_on": {"type": "string","format": "date"},"ends_on": {"type": ["string","null"],"format": "date"},"signed_at": {"type": ["string","null"],"format": "date-time"},"retainer_minutes_per_period": {"type": ["integer","null"],"minimum": 0},"retainer_minutes_per_month": {"type": ["integer","null"],"minimum": 0},"retainer_amount_per_period": {"type": ["integer","null"],"minimum": 0},"hourly_rate_amount": {"type": ["integer","null"],"minimum": 0},"rollover_months": {"type": ["integer","null"],"minimum": 0,"maximum": 120},"project": {"type": ["string","null"],"maxLength": 255},"client_id": {"type": "string","format": "uuid"},"client_name": {"type": "string","maxLength": 160},"project_id": {"type": ["string","null"],"format": "uuid"}}},
      "ClientCompany": {"type": "object","additionalProperties": false,"required": ["version","id","name","billing_email","is_active","automatic_invoice_email_enabled","automatic_invoice_email_delay_days"],"properties": {"version": {"type": "string","minLength": 64,"maxLength": 64},"id": {"type": "string","format": "uuid"},"name": {"type": "string","maxLength": 160},"billing_email": {"type": ["string","null"],"maxLength": 255},"is_active": {"type": "boolean"},"automatic_invoice_email_enabled": {"type": "boolean"},"automatic_invoice_email_delay_days": {"type": ["integer","null"],"minimum": 0,"maximum": 365}}},
...
      "Error": {"type": "object", "additionalProperties": false, "required": ["message"], "properties": {"message": {"type": "string"}, "errors": {"type": "object", "additionalProperties": {"type": "array", "items": {"type": "string"}}}}},
...
    }
  }
}